# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Auction System",
        "description": "Auction System API Documentation",
        "version": "2.0.0"
    },
    "paths": {
...
        "/api/v1/status/product/{product_id}": {
            "put": {
                "tags": [
                    "Status",
                    "Status Management"
                ],
                "summary": "Update Product Status",
                "description": "Update product shipping status (UC02 - Admin only)\n\nPUT /status/product/{product_id}\nHeaders: Authorization: Bearer \u003Caccess_token\u003E\nBody: { \"shipping_status\": \"shipped\" }\nReturns: Updated product information",
                "operationId": "update_product_status_api_v1_status_product__product_id__put",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Product Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductStatusUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": true,
                                    "title": "Response Update Product Status Api V1 Status Product  Product Id  Put"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "Status",
                    "Status Management"
                ],
                "summary": "Get Product Status",
                "description": "View product status after auction (UC04)\n\nGET /status/product/{product_id}\nHeaders: Authorization: Bearer \u003Caccess_token\u003E\nReturns: Product status information with formatted data",
                "operationId": "get_product_status_api_v1_status_product__product_id__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Product Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": true,
                                    "title": "Response Get Product Status Api V1 Status Product  Product Id  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/status/auction/{auction_id}/result": {
            "put": {
                "tags": [
                    "Status",
                    "Status Management"
                ],
                "summary": "Update Auction Result",
                "description": "Update auction result (UC03 - Admin only)\n\nPUT /status/auction/{auction_id}/result\nHeaders: Authorization: Bearer \u003Caccess_token\u003E\nBody: { \"bid_winner_id\": 123 }\nReturns: Updated auction information",
                "operationId": "update_auction_result_api_v1_status_auction__auction_id__result_put",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "auction_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Auction Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuctionResultUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": true,
                                    "title": "Response Update Auction Result Api V1 Status Auction  Auction Id  Result Put"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/status/auction/{auction_id}/complete": {
            "get": {
                "tags": [
                    "Status",
                    "Status Management"
                ],
                "summary": "Get Auction Completion Status",
                "description": "Get comprehensive auction completion status (Admin and winner only)\n\nGET /status/auction/{auction_id}/complete\nHeaders: Authorization: Bearer \u003Caccess_token\u003E\nReturns: Comprehensive auction status with formatted data",
                "operationId": "get_auction_completion_status_api_v1_status_auction__auction_id__complete_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "auction_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Auction Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": true,
                                    "title": "Response Get Auction Completion Status Api V1 Status Auction  Auction Id  Complete Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/status/auction/{auction_id}/finalize": {
            "post": {
                "tags": [
                    "Status",
                    "Status Management"
                ],
                "summary": "Finalize Auction",
                "description": "Finalize auction and update all related statuses (Admin only)\n\nPOST /status/auction/{auction_id}/finalize\nHeaders: Authorization: Bearer \u003Caccess_token\u003E\nReturns: Success message",
                "operationId": "finalize_auction_api_v1_status_auction__auction_id__finalize_post",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "auction_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Auction Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": true,
                                    "title": "Response Finalize Auction Api V1 Status Auction  Auction Id  Finalize Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AuctionResultUpdate": {
                "properties": {
                    "bidWinnerID": {
                        "type": "integer",
                        "title": "Bidwinnerid"
                    }
                },
                "type": "object",
                "required": [
                    "bidWinnerID"
                ],
                "title": "AuctionResultUpdate"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ProductStatusUpdate": {
                "properties": {
                    "shippingStatus": {
                        "type": "string",
                        "title": "Shippingstatus"
                    }
                },
                "type": "object",
                "required": [
                    "shippingStatus"
                ],
                "title": "ProductStatusUpdate"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}